+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/service/cprdomainevents/publishers/AddressEventPublisher.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.personrecord.service.cprdomainevents.listeners
+package uk.gov.justice.digital.hmpps.personrecord.service.cprdomainevents.publishers
 
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
